Better Corridor Spawning
