+++ src/main/java/model/map/object/loot/Chest.java
@@ -1,6 +1,5 @@
 package model.map.object.loot;
 
-import javafx.scene.image.Image;
 import model.map.SeededRandom;
 import model.map.object.IPosition;
 import model.map.specification.texture.TextureHandler;
@@ -14,7 +13,6 @@ public class Chest extends SimpleLootObject {
 
     @Override
     public void draw(IMapCanvas canvas, TextureHandler textureHandler) {
-        Image image = new Image("Icon.png");
-        canvas.drawImage(image, parent, offsetX, offsetY, offfsetZ, 1, 1, rotation);
+        canvas.drawImage(textureHandler.getChest(), parent, offsetX, offsetY, offfsetZ, 1, 1, 0);
     }
 }
